fix: changement de qlq int en int16_t

+++ src/Quake/Q_crandom.c
@@ -1,6 +1,6 @@
+#include <stdint.h> // int16_t
 
-
-float Q_crandom( int *seed )
+float Q_crandom( int16_t *seed )
 {
 	return ( 2.0 * ( Q_random( seed ) - 0.5 ) );
 }
